requirements.txt
alembic==1.4.2
bcrypt==3.2.0
cffi==1.14.2
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-HTTPAuth==4.1.0
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
pycparser==2.20
pyodbc==4.0.30
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
